# api-specification.yaml
openapi: 3.0.0
info:
  title: Distributed Assessment Platform API
  version: 1.0.0
  description: |
    Scalable distributed assessment platform with automated notifications, reporting, and data export capabilities.
    
    ## Features
    - Role-based access control (Admin/Student)
    - Hierarchical content organization (Subject→Chapter→Quiz→Question)
    - Real-time quiz attempts with automatic scoring
    - Automated background jobs for daily reminders and monthly reports
    - Data export capabilities with CSV generation
    
    ## Authentication
    All endpoints require JWT Bearer token authentication except for login and registration.
    
  contact:
    name: Distributed Assessment Platform Support
    email: support@assessmentplatform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.assessmentplatform.com
    description: Production server
...
paths:
...
  /admin/subject:
    get:
      tags: [Admin - Subject]
      summary: Get all subjects
      description: Retrieve a list of all subjects in the system
      responses:
        '200':
          description: List of subjects retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags: [Admin - Subject]
      summary: Create a new subject
      description: Create a new subject in the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectRequest'
            example:
              name: Mathematics
              description: Math subject
      responses:
        '201':
          description: Subject created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /admin/subject/{subject_id}:
    get:
      tags: [Admin - Subject]
      summary: Get subject by ID
      description: Retrieve a specific subject by its ID
      parameters:
        - name: subject_id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the subject to retrieve
      responses:
        '200':
          description: Subject details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags: [Admin - Subject]
      summary: Update subject
      description: Update an existing subject
      parameters:
        - name: subject_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectRequest'
            example:
              name: Mathematics
              description: Updated description
      responses:
        '200':
          description: Subject updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [Admin - Subject]
      summary: Delete subject
      description: Delete a subject from the system
      parameters:
        - name: subject_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Subject deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'
  /admin/subject/search:
    get:
      tags: [Admin - Subject]
      summary: Search subjects
      description: Search for subjects by name or description
      parameters:
        - name: text
          in: query
          required: true
          schema:
            type: string
            minLength: 3
          description: Search text (minimum 3 characters)
      responses:
        '200':
          description: Subject search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectListResponse'

  # Admin - Chapter Management
...
  /auth/login:
    post:
      tags: [Auth]
      summary: User login
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: admin@assessmentplatform.com
              password: yourpassword
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    SubjectRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 512
        description:
          type: string
          maxLength: 1024
    SubjectResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SubjectListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/SubjectResponse'

    # Chapter Schemas
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 120
        password:
          type: string
          minLength: 6
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            access_token:
              type: string
            user:
              $ref: '#/components/schemas/UserInfo'
...
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        fullname:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, student]
        status:
          type: string
          enum: [active, inactive, suspended]
        last_login:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error_message:
          type: string
        code:
          type: integer
  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error_message: "Invalid input data"
            code: 400
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error_message: "Authentication required"
            code: 401
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error_message: "Insufficient permissions"
            code: 403
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error_message: "Resource not found"
            code: 404
security:
  - BearerAuth: []